post- primera revision

+++ AluminiosRuta5/Objects/Perfil.cs
@@ -11,9 +11,9 @@ namespace AluminiosRuta5.Objects
         public int PerfilId { get; set; }
         public string Codigo { get; set; }
         public string Descripcion { get; set; }
-        public double KgXPaquete { get; set; }
+        public decimal KgXPaquete { get; set; }
         public int CantidadTiras { get; set; }
-        public double Import { get; set; }
+        public decimal Import { get; set; }
 
         public int CategoriaId { get; set; }
